[MAINTENANCE] Move type_comparison module to expectations/ package level
The module serves both expectations/core/ and expectations/metrics/
equally — it's shared infrastructure, not specific to core. Moving it
to expectations/type_comparison.py gives a cleaner import path and
avoids metrics importing from expectations.core.
